requirements.txt
anthropic
pyairtable
watchdog
python-telegram-bot>=20.0
python-dotenv
httpx
tenacity
requests==2.31.0
solana
cryptography
base58
